Deduplicate between title_bar and collab modal
