+++ crates/call/src/call_impl/mod.rs
@@ -1,3 +1,4 @@
+pub mod diagnostics;
 pub mod participant;
 pub mod room;
