+++ crates/title_bar/src/collab.rs
@@ -1,6 +1,5 @@
 use std::rc::Rc;
 use std::sync::Arc;
-use std::time::Duration;
 
 use call::{ActiveCall, Room};
 use channel::ChannelStore;
@@ -25,208 +24,6 @@ use workspace::{ParticipantLocation, notifications::DetachAndPromptErr};
 use zed_actions::ShowCallStats;
 
 use crate::TitleBar;
-
-#[derive(Clone, Default)]
-pub struct CallStats {
-    pub latency_ms: Option<f64>,
-    pub jitter_ms: Option<f64>,
-    pub packet_loss_pct: Option<f64>,
-    pub input_lag: Option<Duration>,
-}
-
-impl TitleBar {
-    pub(crate) fn start_call_stats_polling(&mut self, cx: &mut gpui::Context<Self>) {
-        self.call_stats_poll_task = Some(cx.spawn(async move |this, cx| {
-            loop {
-                if this
-                    .update(cx, |this, cx| this.poll_call_stats(cx))
-                    .is_err()
-                {
-                    break;
-                }
-                cx.background_executor().timer(Duration::from_secs(1)).await;
-            }
-        }));
-    }
-
-    pub(crate) fn stop_call_stats_polling(&mut self) {
-        self.call_stats_poll_task.take();
-        self.call_stats = CallStats::default();
-    }
-
-    fn poll_call_stats(&mut self, cx: &mut gpui::Context<Self>) {
-        let Some(room) = ActiveCall::global(cx).read(cx).room().cloned() else {
-            return;
-        };
-
-        self.call_stats.input_lag = room.read(cx).input_lag();
-        let stats_future = room.read(cx).get_stats(cx);
-
-        let background_task = cx.background_executor().spawn(async move {
-            let session_stats = stats_future.await;
-            session_stats.map(|stats| compute_network_stats(&stats))
-        });
-
-        self.call_stats_update_task = Some(cx.spawn(async move |this, cx| {
-            let result = background_task.await;
-            this.update(cx, |this, cx| {
-                if let Some(computed) = result {
-                    this.call_stats.latency_ms = computed.latency_ms;
-                    this.call_stats.jitter_ms = computed.jitter_ms;
-                    this.call_stats.packet_loss_pct = computed.packet_loss_pct;
-                }
-                cx.notify();
-            })
-            .ok();
-        }));
-    }
-}
-
-struct ComputedNetworkStats {
-    latency_ms: Option<f64>,
-    jitter_ms: Option<f64>,
-    packet_loss_pct: Option<f64>,
-}
-
-fn compute_network_stats(stats: &livekit_client::SessionStats) -> ComputedNetworkStats {
-    let mut min_rtt: Option<f64> = None;
-    let mut max_jitter: Option<f64> = None;
-    let mut total_packets_received: u64 = 0;
-    let mut total_packets_lost: i64 = 0;
-
-    let all_stats = stats
-        .publisher_stats
-        .iter()
-        .chain(stats.subscriber_stats.iter());
-
-    for stat in all_stats {
-        extract_metrics(
-            stat,
-            &mut min_rtt,
-            &mut max_jitter,
-            &mut total_packets_received,
-            &mut total_packets_lost,
-        );
-    }
-
-    let total_expected = total_packets_received as i64 + total_packets_lost;
-    let packet_loss_pct = if total_expected > 0 {
-        Some((total_packets_lost as f64 / total_expected as f64) * 100.0)
-    } else {
-        None
-    };
-
-    ComputedNetworkStats {
-        latency_ms: min_rtt.map(|rtt| rtt * 1000.0),
-        jitter_ms: max_jitter.map(|j| j * 1000.0),
-        packet_loss_pct,
-    }
-}
-
-#[cfg(all(
-    not(rust_analyzer),
-    any(
-        test,
-        feature = "test-support",
-        all(target_os = "windows", target_env = "gnu"),
-        target_os = "freebsd"
-    )
-))]
-fn extract_metrics(
-    _stat: &livekit_client::RtcStats,
-    _min_rtt: &mut Option<f64>,
-    _max_jitter: &mut Option<f64>,
-    _total_packets_received: &mut u64,
-    _total_packets_lost: &mut i64,
-) {
-}
-
-#[cfg(any(
-    rust_analyzer,
-    not(any(
-        test,
-        feature = "test-support",
-        all(target_os = "windows", target_env = "gnu"),
-        target_os = "freebsd"
-    ))
-))]
-fn extract_metrics(
-    stat: &livekit_client::RtcStats,
-    min_rtt: &mut Option<f64>,
-    max_jitter: &mut Option<f64>,
-    total_packets_received: &mut u64,
-    total_packets_lost: &mut i64,
-) {
-    use livekit_client::RtcStats;
-
-    match stat {
-        RtcStats::CandidatePair(pair) => {
-            let rtt = pair.candidate_pair.current_round_trip_time;
-            if rtt > 0.0 {
-                *min_rtt = Some(match *min_rtt {
-                    Some(current) => current.min(rtt),
-                    None => rtt,
-                });
-            }
-        }
-        RtcStats::InboundRtp(inbound) => {
-            let jitter = inbound.received.jitter;
-            if jitter > 0.0 {
-                *max_jitter = Some(match *max_jitter {
-                    Some(current) => current.max(jitter),
-                    None => jitter,
-                });
-            }
-            *total_packets_received += inbound.received.packets_received;
-            *total_packets_lost += inbound.received.packets_lost;
-        }
-        RtcStats::RemoteInboundRtp(remote_inbound) => {
-            let rtt = remote_inbound.remote_inbound.round_trip_time;
-            if rtt > 0.0 {
-                *min_rtt = Some(match *min_rtt {
-                    Some(current) => current.min(rtt),
-                    None => rtt,
-                });
-            }
-        }
-        _ => {}
-    }
-}
-
-fn metric_quality(value: f64, warn_threshold: f64, error_threshold: f64) -> ConnectionQuality {
-    if value < warn_threshold {
-        ConnectionQuality::Excellent
-    } else if value < error_threshold {
-        ConnectionQuality::Poor
-    } else {
-        ConnectionQuality::Lost
-    }
-}
-
-/// Computes the effective connection quality by taking the worst of the
-/// LiveKit-reported quality and each individual metric rating.
-fn effective_connection_quality(
-    livekit_quality: ConnectionQuality,
-    stats: &CallStats,
-) -> ConnectionQuality {
-    let mut worst = livekit_quality;
-
-    if let Some(latency) = stats.latency_ms {
-        worst = worst.max(metric_quality(latency, 100.0, 300.0));
-    }
-    if let Some(jitter) = stats.jitter_ms {
-        worst = worst.max(metric_quality(jitter, 30.0, 75.0));
-    }
-    if let Some(loss) = stats.packet_loss_pct {
-        worst = worst.max(metric_quality(loss, 1.0, 5.0));
-    }
-    if let Some(lag) = stats.input_lag {
-        let lag_ms = lag.as_secs_f64() * 1000.0;
-        worst = worst.max(metric_quality(lag_ms, 20.0, 50.0));
-    }
-
-    worst
-}
 
 fn format_stat(value: Option<f64>, format: impl Fn(f64) -> String) -> String {
     match value {
@@ -559,7 +356,11 @@ impl TitleBar {
         let can_use_microphone = room.can_use_microphone();
         let can_share_projects = room.can_share_projects();
         let screen_sharing_supported = cx.is_screen_capture_supported();
-        let connection_quality = room.connection_quality();
+
+        let stats = room
+            .diagnostics()
+            .map(|d| d.read(cx).stats().clone())
+            .unwrap_or_default();
 
         let channel_store = ChannelStore::global(cx);
         let channel = room
@@ -568,7 +369,7 @@ impl TitleBar {
 
         let mut children = Vec::new();
 
-        let effective_quality = effective_connection_quality(connection_quality, &self.call_stats);
+        let effective_quality = stats.effective_quality.unwrap_or(ConnectionQuality::Lost);
         let (signal_icon, signal_color, quality_label) = match effective_quality {
             ConnectionQuality::Excellent => {
                 (IconName::SignalHigh, Some(Color::Success), "Excellent")
@@ -577,8 +378,6 @@ impl TitleBar {
             ConnectionQuality::Poor => (IconName::SignalMedium, Some(Color::Warning), "Poor"),
             ConnectionQuality::Lost => (IconName::SignalLow, Some(Color::Error), "Lost"),
         };
-
-        let stats = self.call_stats.clone();
         let quality_label: SharedString = quality_label.into();
         children.push(
             IconButton::new("call-quality", signal_icon)
