+++ crates/title_bar/src/title_bar.rs
@@ -158,9 +158,7 @@ pub struct TitleBar {
     banner: Entity<OnboardingBanner>,
     update_version: Entity<UpdateVersion>,
     screen_share_popover_handle: PopoverMenuHandle<ContextMenu>,
-    call_stats: collab::CallStats,
-    call_stats_poll_task: Option<gpui::Task<()>>,
-    call_stats_update_task: Option<gpui::Task<()>>,
+    _diagnostics_subscription: Option<gpui::Subscription>,
 }
 
 impl Render for TitleBar {
@@ -415,15 +413,10 @@ impl TitleBar {
             banner,
             update_version,
             screen_share_popover_handle: PopoverMenuHandle::default(),
-            call_stats: collab::CallStats::default(),
-            call_stats_poll_task: None,
-            call_stats_update_task: None,
+            _diagnostics_subscription: None,
         };
 
-        // Start polling if already in a call.
-        if ActiveCall::global(cx).read(cx).room().is_some() {
-            this.start_call_stats_polling(cx);
-        }
+        this.observe_diagnostics(cx);
 
         this
     }
@@ -969,16 +962,21 @@ impl TitleBar {
     }
 
     fn active_call_changed(&mut self, cx: &mut Context<Self>) {
-        let in_call = ActiveCall::global(cx).read(cx).room().is_some();
-        let was_polling = self.call_stats_poll_task.is_some();
-
-        if in_call && !was_polling {
-            self.start_call_stats_polling(cx);
-        } else if !in_call && was_polling {
-            self.stop_call_stats_polling();
-        }
-
+        self.observe_diagnostics(cx);
         cx.notify();
+    }
+
+    fn observe_diagnostics(&mut self, cx: &mut Context<Self>) {
+        let diagnostics = ActiveCall::global(cx)
+            .read(cx)
+            .room()
+            .and_then(|room| room.read(cx).diagnostics().cloned());
+
+        if let Some(diagnostics) = diagnostics {
+            self._diagnostics_subscription = Some(cx.observe(&diagnostics, |_, _, cx| cx.notify()));
+        } else {
+            self._diagnostics_subscription = None;
+        }
     }
 
     fn share_project(&mut self, cx: &mut Context<Self>) {
